8219139: move hotspot tests from test/jdk/vm Reviewed-by: dholmes, mseledtsov
