+++ test/hotspot/jtreg/compiler/codegen/ShiftTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -25,7 +25,11 @@
  * @test
  * @bug 4093292
  * @summary Test for correct code generation by the JIT
+ * @run main compiler.codegen.ShiftTest
  */
+
+package compiler.codegen;
+
 public class ShiftTest {
     static final int w = 32;
