+++ test/hotspot/jtreg/compiler/exceptions/ExceptionInInit.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018, 2019, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -29,7 +29,10 @@
  * @author Tom Rodriguez
  *
  * @modules java.rmi
+ * @run main compiler.exceptions.ExceptionInInit
  */
+
+package compiler.exceptions;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
