+++ test/jdk/tools/launcher/JniInvocationTest.java
